Refactor all login/logout/current user code into UILogin.
